fix:fixed some bugs in backend and frontend component

+++ Backend/src/main/java/com/swa/filter/Repository/FolderDirRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.swa.filter.mySQLTables.FolderDir;
 
-public interface FolderDirRepository extends JpaRepository<FolderDir,Long>{
+public interface FolderDirRepository extends JpaRepository<FolderDir,Integer>{
     Optional<FolderDir>findByName(String name);
     Optional<FolderDir>findByPath(String name);
 }
